+++ HISTORY.md
@@ -18,6 +18,7 @@
 * objects: adapted formula editor for variadic infix reporters
 * objects: removed scaffolding
 * store: apply block migration dictionary to hidden blocks in microworlds
+* store: removed scaffolding
 
 ### 2022-02-28
 * blocks, objects, threads, store: made addition reporter variadic